+++ test/replay/driver.go
@@ -7,6 +7,12 @@ import (
 	"encoding/json"
 	"fmt"
 
+<<<<<<< HEAD
+=======
+	appparams "github.com/babylonlabs-io/babylon/v3/app/params"
+	"github.com/babylonlabs-io/babylon/v3/app/signingcontext"
+
+>>>>>>> 2b02d75 (Implement context separator signing (#1252))
 	"math/rand"
 	"path/filepath"
 	"testing"
@@ -1020,4 +1026,20 @@ func (d *BabylonAppDriver) CreateFinalityProviderForConsumer(consumer *Consumer)
 	fp.RegisterFinalityProvider(consumer.ID)
 
 	return fp
+}
+
+func (d *BabylonAppDriver) FpPopContext() string {
+	return signingcontext.FpPopContextV0(d.App.ChainID(), d.App.BTCStakingKeeper.ModuleAddress())
+}
+
+func (d *BabylonAppDriver) StakerPopContext() string {
+	return signingcontext.StakerPopContextV0(d.App.ChainID(), d.App.BTCStakingKeeper.ModuleAddress())
+}
+
+func (d *BabylonAppDriver) FpRandCommitContext() string {
+	return signingcontext.FpRandCommitContextV0(d.App.ChainID(), d.App.FinalityKeeper.ModuleAddress())
+}
+
+func (d *BabylonAppDriver) FpFinVoteContext() string {
+	return signingcontext.FpFinVoteContextV0(d.App.ChainID(), d.App.FinalityKeeper.ModuleAddress())
 }