+++ CHANGELOG.md
@@ -104,6 +104,12 @@ when the babylon block is being BTC rewarded.
 - [#525](https://github.com/babylonlabs-io/babylon/pull/525) fix: add back `NewIBCHeaderDecorator` post handler
 - [#964](https://github.com/babylonlabs-io/babylon/pull/964) fix: stateless validation `ValidateBasic` of
 `MsgWrappedCreateValidator` to avoid panic in transaction submission
+<<<<<<< HEAD
+=======
+- [#1247](https://github.com/babylonlabs-io/babylon/pull/1247) fix: zoneconcierge ibc router
+- [#1252](https://github.com/babylonlabs-io/babylon/pull/1252) Implement context separation
+between different signing operations
+>>>>>>> 2b02d75 (Implement context separator signing (#1252))
 
 ## v2.1.0